requirements.txt
alembic==1.1.0
appdirs==1.4.3
arrow==0.15.4
astroid==1.6.6
attrs==19.1.0
autokey==0.95.7
autopep8==1.4.4
bcrypt==3.1.7
blinker==1.4
blivet==3.1.4
blivet-gui==2.1.10
Brlapi==0.6.7
CairoSVG==1.0.20
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cupshelpers==1.0
Django==2.2.7
Faker==2.0.1
Flask==1.1.1
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-Compress==1.4.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RQ==0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
gpg==1.12.0
gunicorn==19.9.0
honcho==1.0.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsmin==2.2.2
jsonpickle==1.2
langtable==0.0.43
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
Naked==0.1.31
notify2==0.3.1
ntplib==0.3.3
packaging==19.1
paytm-pg==1.0.6
pexpect==4.6.0
pipupgrade==1.5.1
psycopg2==2.8.3
ptyprocess==0.6.0
pwquality==1.4.0
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.74
pycurl==7.43.0.2
PyGObject==3.32.2
pyinotify==0.9.6
pykickstart==3.20
pylint==1.9.4
pyparsing==2.4.2
pyparted==3.11.2
python-augeas==0.5.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-meh==0.47
python-xapp==1.6.0
python-xlib==0.25
pytz==2019.3
pyxdg==0.26
PyYAML==5.1
raygun4py==4.3.0
redis==3.3.11
requests==2.22.0
rope==0.14.0
rpm==4.14.2.1
rq==1.1.0
rq-dashboard==0.6.0
selinux==2.9
sepolicy==1.1
setroubleshoot==1.1
shellescape==3.4.1
simpleline==1.4
six==1.13.0
slip==0.6.4
slip.dbus==0.6.4
sos==3.7
SQLAlchemy==1.3.11
sqlparse==0.3.0
systemd-python==234
text-unidecode==1.2
timeloop==1.0.2
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.8
webassets==0.12.1
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
